Introduce `JarUnsigner`
